\name{epmrob-package}
\alias{epmrob-package}
%\alias{epmrob}
\docType{package}
\title{
Robust estimation of probit models with endogeneity
}
\description{
Package provides a set of tools for robust estimation of probit models with endogeneity.
}
\details{
\tabular{ll}{
Package: \tab epmrob \cr
Type: \tab Package \cr
Version: \tab 1.0 \cr
Date: \tab 2023-06-14 \cr
License: \tab GPL-2 \cr
}
}
\author{
Mikhail Zhelonkin, Andre Bik, Andrea Naghi

Maintainer: Mikhail Zhelonkin <Mikhail.Zhelonkin@gmail.com>
}
\references{
Naghi, A. A., Varadi, M., & Zhelonkin, M. (2022). Robust Estimation of Probit Models with Endogeneity. \emph{Econometrics and Statistics}. \doi{10.1016/j.ecosta.2022.05.001}
}
\keyword{ Robust estimation }
\keyword{ Probit model }
\keyword{ Endogenous variable }
\seealso{
\code{\link{epmrob}}, \code{\link{IVProbitRob}}
}
