\name{print.epmrob}
\alias{print.epmrob}
\title{
Print a \code{epmrob} object
}
\description{
Print an object generated by \code{\link{epmrob}}.
}
\usage{
\method{print}{epmrob}(x, digits = 4, ...)
}
\arguments{
  \item{x}{
object of \code{epmrob} class.
}
  \item{digits}{
number of significant digits to be printed.
}
  \item{\dots}{
currently not used.
}
}
\value{
No return value.
}
\author{
Mikhail Zhelonkin, Andre Bik, Andrea Naghi
}
\seealso{
\code{\link{epmrob}}
}
