\name{print.summary.epmrob}
\alias{print.summary.epmrob}
\title{
Print Function for \code{summary.epmrob}
}
\description{
Print a \code{\link{summary.epmrob}} object.
}
\usage{
\method{print}{summary.epmrob}(x, digits = 4, ...)
}
\arguments{
  \item{x}{
object of class \code{summary.epmrob} returned by a summary function.
}
  \item{digits}{
number of nonzero digits after comma.
}
  \item{\dots}{
currently not used.
}
}
\value{
No return value.
}
\author{
Mikhail Zhelonkin, Andre Bik, Andrea Naghi
}