\name{summary.epmrob}
\alias{summary.epmrob}
\title{
Summarizing Robust Fits of Endogenous Probit Models
}
\description{
Summarizes robust fit of endogenous probit models.
}
\usage{
\method{summary}{epmrob}(object, ...)
}
\arguments{
  \item{object}{
object of class "\code{\link{epmrob}}".
}
  \item{\dots}{
currently not used.
}
}
\value{
object of class "\code{summary.epmrob}".
}
\author{
Mikhail Zhelonkin, Andre Bik, Andrea Naghi
}
\seealso{
\code{\link{epmrob}}
}